Rename and move BucketSettings model
